refactor(frontend): move chat composer state to shared feature
