+++ codemap.md
@@ -89,8 +89,8 @@ OpenDucktor is a Bun monorepo for a macOS-first Tauri v2 desktop app and local b
 | `packages/frontend/src/components/` | Reusable UI and feature widgets, including layout chrome, shadcn primitives, error screens, markdown/document preview controls, and feature panels/modals. | [packages/frontend/src/components/codemap.md](packages/frontend/src/components/codemap.md) |
 | `packages/frontend/src/components/errors/` | Crash and fatal-error handling for the frontend app. | [packages/frontend/src/components/errors/codemap.md](packages/frontend/src/components/errors/codemap.md) |
 | `packages/frontend/src/components/features/` | Feature UI blocks for Kanban, task details/composer, repository switching, settings, diagnostics, PRs, and Agent Studio panels. | [packages/frontend/src/components/features/codemap.md](packages/frontend/src/components/features/codemap.md) |
-| `packages/frontend/src/components/features/agents/` | Agent Studio feature UI, including chat, session tabs, workspace sidebar, dev-server panel, git panel, runtime attachment retry handling, and session-start entry points. | [packages/frontend/src/components/features/agents/codemap.md](packages/frontend/src/components/features/agents/codemap.md) |
-| `packages/frontend/src/components/features/agents/agent-chat/` | Agent Studio chat surface: transcript thread, composer, tool cards, permissions/questions, session todos, runtime attachment retry, document preview/copy controls, and transcript dialog. | [packages/frontend/src/components/features/agents/agent-chat/codemap.md](packages/frontend/src/components/features/agents/agent-chat/codemap.md) |
+| `packages/frontend/src/components/features/agents/` | Agent feature UI, including reusable chat, Agent Studio session tabs, workspace sidebar, dev-server panel, git panel, runtime attachment retry handling, and session-start entry points. | [packages/frontend/src/components/features/agents/codemap.md](packages/frontend/src/components/features/agents/codemap.md) |
+| `packages/frontend/src/components/features/agents/agent-chat/` | Reusable agent chat surface: transcript thread, composer, tool cards, permissions/questions, session todos, runtime attachment retry, document preview/copy controls, and transcript dialog. | [packages/frontend/src/components/features/agents/agent-chat/codemap.md](packages/frontend/src/components/features/agents/agent-chat/codemap.md) |
 | `packages/frontend/src/components/features/agents/agent-studio-git-panel/` | Agent Studio git/worktree UI for diffs, commit composer, open-in tooling, and push/rebase confirmation flows. | [packages/frontend/src/components/features/agents/agent-studio-git-panel/codemap.md](packages/frontend/src/components/features/agents/agent-studio-git-panel/codemap.md) |
 | `packages/frontend/src/components/features/diagnostics/` | Health/status diagnostics UI for workspace, Beads, and runtime checks. | [packages/frontend/src/components/features/diagnostics/codemap.md](packages/frontend/src/components/features/diagnostics/codemap.md) |
 | `packages/frontend/src/components/features/kanban/` | Kanban board renderers, task cards, workflow controls, virtualization, and status/theme presentation. | [packages/frontend/src/components/features/kanban/codemap.md](packages/frontend/src/components/features/kanban/codemap.md) |
@@ -106,6 +106,7 @@ OpenDucktor is a Bun monorepo for a macOS-first Tauri v2 desktop app and local b
 | `packages/frontend/src/components/ui/` | Shared shadcn-style primitives, markdown renderer pieces, document copy/preview affordances, and UI helpers. | [packages/frontend/src/components/ui/codemap.md](packages/frontend/src/components/ui/codemap.md) |
 | `packages/frontend/src/contexts/` | Cross-tree React context providers that sit outside the app-state provider graph. | [packages/frontend/src/contexts/codemap.md](packages/frontend/src/contexts/codemap.md) |
 | `packages/frontend/src/features/` | Cross-cutting workflows above page components and below app-wide state. | [packages/frontend/src/features/codemap.md](packages/frontend/src/features/codemap.md) |
+| `packages/frontend/src/features/agent-chat-composer/` | Reusable agent chat composer state helpers for session context, model selection, prompt input, file search, slash commands, and context usage. | [packages/frontend/src/features/agent-chat-composer/codemap.md](packages/frontend/src/features/agent-chat-composer/codemap.md) |
 | `packages/frontend/src/features/agent-studio-build-tools/` | Build-tools diff/dev-server read models and bootstrap helpers for Agent Studio. | [packages/frontend/src/features/agent-studio-build-tools/codemap.md](packages/frontend/src/features/agent-studio-build-tools/codemap.md) |
 | `packages/frontend/src/features/agent-studio-git/` | Agent Studio git diff state, polling, refresh, and worktree resolution helpers. | [packages/frontend/src/features/agent-studio-git/codemap.md](packages/frontend/src/features/agent-studio-git/codemap.md) |
 | `packages/frontend/src/features/autopilot/` | Autopilot rule/event catalog for task-state-driven workflow launches. | [packages/frontend/src/features/autopilot/codemap.md](packages/frontend/src/features/autopilot/codemap.md) |
