+++ packages/frontend/src/features/agent-chat-composer/model-selection/use-model-selection-actions.ts
