+++ packages/frontend/src/features/agent-chat-composer/prompt-input/create-chat-composer-file-search.ts
