+++ packages/frontend/src/features/agent-chat-composer/prompt-input/use-chat-composer-slash-commands.test.ts
