+++ packages/frontend/src/features/agent-chat-composer/session-context/codemap.md
@@ -1,0 +1,10 @@
+# packages/frontend/src/features/agent-chat-composer/session-context/
+
+## Responsibility
+Reusable active-session context extraction for agent chat composers: session id, repo path, runtime kind, working directory, catalog loading state, live context usage, and message history.
+
+## Design Patterns
+Keep this folder about session-derived composer inputs only. Do not place model-selection preference, prompt-input, or context-usage calculation logic here.
+
+## Integration Points
+Consumed by page-level composer adapters before model-selection, prompt-input, and context-usage helpers run.
